requirements.txt
# --- Core Framework ---
fastapi==0.111.0
uvicorn==0.30.1

# --- Database & ORM ---
sqlalchemy==2.0.31
sqlmodel==0.0.14
aiomysql==0.2.0  # Основний асинхронний драйвер
pymysql==1.1.1   # Синхронна залежність aiomysql

# --- Authentication & Security ---
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
argon2-cffi # Версія має бути закріплена
cryptography==42.0.8

# --- Networking ---
pywinrm==0.4.3
ldap3==2.9.1
requests-ntlm==1.2.0
ntlm-auth==1.5.0

# --- Pydantic & Settings ---
pydantic==2.8.2
pydantic-settings==2.3.3